prime number code

+++ Core/src/core/practical/cases/ArmstrongNumberChecker.java
@@ -3,7 +3,7 @@ package core.practical.cases;
 import java.util.Scanner;
 
 public class ArmstrongNumberChecker {
-	private static int findDigits(int input) {
+	private static int findDigits(long input) {
 		int count = 0;
 		while (input > 0) {
 			input /= 10;
@@ -14,8 +14,8 @@ public class ArmstrongNumberChecker {
 
 	}
 
-	private static void ArmstrongNumber(int originalNumber) {
-		int remainder = 0, result = 0, temp = originalNumber;
+	private static void ArmstrongNumber(long originalNumber) {
+		long remainder = 0, result = 0, temp = originalNumber;
 		int totalDigits = findDigits(originalNumber);
 		while (temp > 0) {
 			remainder = temp % 10;
@@ -33,7 +33,7 @@ public class ArmstrongNumberChecker {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		System.out.println("kindly enter the input :");
-		int originalNumber = scan.nextInt();
+		long originalNumber = scan.nextLong();
 		ArmstrongNumber(originalNumber);
 		scan.close();
 	}
